requirements.txt
discord.py
requests
pymongo[srv]
flask
gunicorn